#79 code-review and tests
* #79 code-review

* #79 Multi-line and comment support: removed unnecessary checking, created test class
PR #100 by alexlust

* #79 fixed merge problems

* #79 whitespace fixed

* #79 whitespace fixed

* #79 moved to util package
We don't need a new package. I'm not sure if util is right or the
calling package org.jxls.builder.xls would be better?

* #79 moved to util package
We don't need a new package. I'm not sure if util is right or the
calling package org.jxls.builder.xls would be better?

* #79 more tests, LiteralsExtractor bugfixes